refactor(runner): split start heartbeat module
